requirements.txt
Flask
pandas
openpyxl
google-generativeai
